move autobahn.xbr._util.hl* to autobahn.util and check for click at run-time
