+++ autobahn/twisted/xbr/__init__.py
@@ -43,7 +43,7 @@ if HAS_XBR:
 
     import uuid
 
-    from autobahn.xbr._util import hl
+    from autobahn.util import hl
     from autobahn.xbr._interfaces import IProvider, ISeller, IConsumer, IBuyer, IDelegate
     from autobahn.xbr import _seller, _buyer, _blockchain
